// src/app/model/debug-level.ts
export class CreateDebugLevel {
    ApexCode: string;
    ApexProfiling: string;
    Callout: string;
    Database: string;
    DeveloperName: string;
    MasterLabel: string;
    System: string;
    Validation: string;
    Visualforce: string;
    Workflow: string;

    constructor() {
        this.ApexCode = "";
        this.ApexProfiling = "";
        this.Callout = "";
        this.Database = "";
        this.DeveloperName = "";
        this.MasterLabel = "";
        this.System = "";
        this.Validation = "";
        this.Visualforce = "";
        this.Workflow = "";

    }
}